Rename frame to display
