+++ client/src/components/Visualizations/VisualizationDisplay.vue
