+++ client/src/entry/analysis/router.js
@@ -38,7 +38,7 @@ import UserPreferences from "components/User/UserPreferences";
 import UserPreferencesForm from "components/User/UserPreferencesForm";
 import VisualizationsList from "components/Visualizations/Index";
 import VisualizationCreate from "components/Visualizations/VisualizationCreate";
-import VisualizationFrame from "components/Visualizations/VisualizationFrame";
+import VisualizationDisplay from "components/Visualizations/VisualizationDisplay";
 import VisualizationPublished from "components/Visualizations/VisualizationPublished";
 import HistoryInvocations from "components/Workflow/HistoryInvocations";
 import TrsImport from "components/Workflow/Import/TrsImport";
@@ -660,7 +660,7 @@ export function getRouter(Galaxy) {
                     },
                     {
                         path: "visualizations/display",
-                        component: VisualizationFrame,
+                        component: VisualizationDisplay,
                         name: "VisualizationsDisplay",
                         props: (route) => ({
                             datasetId: route.query.dataset_id,
